opkg: Fix arch evaluation bug
This patch fixes a bug when opkg is installing a package with multiple
implementations with the same name and version but different Architecture
values.

Prior to this fix opkg would not consider the priority of the various arch
values when choosing which package to install.

This change was originally written against master and backported to v0.3.4.

Signed-off-by: Ken Sharp <ken.sharp@ni.com>