bblayers.conf.sample: remove sed hackage
I mistakenly thought we can have the perishable build directories in
any filesystem path. That is just not possible because the code doesn't
have any way to detect where the sources are. Revert back to the
original bblayers.conf content and use that for the sample file.

The restriction is that build directories need to live under the
nilrt.git root, it always was that way and the code is cleaner now.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>