u-boot-mkimage: add bbappend
Also package the dumpimage tool which is a companion to mkimage to
extract an itb (mkimage creates it).

I don't have time to refactor this recipe to rename it to
u-boot-tools which creates the ${PN}-mkimage and ${PN}-dumpimage as
this is a P0. I'll do it in the rebase.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>