images: remove niconsole-image-emulated
The old IMAGES_FSTYPES = vmdk option on which this image depends got
dropped from OE Sumo in favor of the wic infrastructure & types.

Nobody is using virtual machines on CG (we support it only on NXG
builds) and instead of converting this dead logic to WIC I'm dropping
the image to cleanup the tree.

When/if we decide to support CG VMs for someone who will actually use
it, we'll directly implement using WIC like we do on NXG.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>