curl: bump 7.53.1 -> 7.61.0
Dropped patch 0005-Added-curl_easy_setopt_va-API-function.patch
because a new function Curl_vsetopt got added upstream which is
identical to what we had in the patch (and updates patch 0009 to use
the upstream function).

Refreshed other patches.

Removed the % versioned bbappend because devtool complains and starts
editing the wrong bbappend if you have two bbappends for the same
recipe in the same layer location. Kept the ENABLE_SRC_INSTALL var in
the main bbappend.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>