grub: bump 2.00 -> 2.02
And significantly clean up the recipe because upstream dropped the
_git version and implemented a proper non-efi grub recipe which we can
fully reuse.

We still keep our own grub-efi recipe for its grub-efi-nilrt package
because we use grubx64.efi for the binary name instead of bootx64.efi
for whatever reason (I'll fix this in the future hopefully) and the
smbios table support patch we keep around.

Because wic's isoimage-isohybrid plugin went through a significant
refactoring in sumo, it now uses the grub-efi package directly without
building it's own grub-efi, so we don't override the GRUB_BUILDIN var
anymore in order to inherit the modules needed to boot iso's.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>