opkg: process "Essential" control field at solve time
When Essential was set to 'yes' for a package, opkg would uninstall
dependencies for Essential packages before accounting that the top-level
package should not be uninstalled.

This patch fixes that issue by passing the Essential meta-data to the solver
so the solver can determine if the entire operation should fail prior to any
individual uninstall operations.

Signed-off-by: Ken Sharp <ken.sharp@ni.com>