tune-cortexa9: delete obsolete file
The history of this file is tragi-comic: It got copied sometime before
2013 in p4 from oe-core then got copied into meta-nilrt in 2013. Since
then it wasn't touched or synced with the evolving oe-core file.

Our machines had the following line which matches BOTH our meta-nilrt
and the oe-core file paths:

require conf/machine/include/tune-cortexa9.inc

The content of the meta-nilrt file is so stale, the variables defined
there aren't used at all, the builds just worked with the oe-core file.

Delete it.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>