ninextos: fix dependencies & remove redundant code
We depend on grub-efi not on plain grub (which is non-efi aka legacy).

A dependency on linux-nilrt:do_deploy was missing, we need to make
sure the kernel is in the img deploy dir.

Also DEPENDS = "restore-mode-image" is redundant since we add
do_install[depends] = "restore-mode-image:do_image_complete".

FILES_${PN} = "" is also redundant because we overwrite the variable
based on machines just below it, no use in assinging "" then appending
to it.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>