provisioning/migration: add safemode-restore-image support
This is the meta-nilrt counterpart commit of the meta-orgconf
backwards x64 migration rearchitecture. The rationale for this
migration rearchitecture is to not depend on so many p4/BS exports,
and to avoid provisioning using a hacked up runmode image when the
restore-mode-image infrastructure is already built in OE and used
accross all migration paths (arm backward and forward + x64 forward).

An added benefit is that all our provisioning code is now in one
place, not split between meta-nilrt and meta-orgconf.

safemode-restore-image provisions an older nilrt safemode using the
NXG recovery. This commits adds support in the recovery's image PID 1
provisioning scripts to actually format & install the safemode via
the ni_provisioning.safemode file.

ni_provisioning.safemode has lived until now in meta-orgconf's nioldos
recipe and was copied way back in 2015-2016 from p4/BS's recoverytool,
so the code in it is somewhat stale. I've added a TODO item to merge
the latest changes from p4 into it.

The provisioning scripts are very ugly and need a lot of cleaning up
and code-sharing that's outside the scope of this refactoring, but
it's also added to my TODO.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>