openssh: enable PasswordAuthentication
If it's set to no, all password-based auth including the default
password-less root auth won't work, ssh exits with the error:
Password authentication is disabled to avoid man-in-the-middle
attacks.

This used to work in previous openssh versions, pre-sumo's 7.6p1, now
it looks like the root account is also tied to the wider pass auth
setting.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>