nilrt.inc: use ipk data gz compression instead of xz
We build & use ipk-packaged (and non-ipk's from sources from OE)
software in nibuild and for these the source "stager" components
inside nibuild need to support the ipk data compression formats of
OE/opkg.

OE migrated to xz compression in Sumo and nibuild's stager does not
support it yet, so revert back to gz until it's supported.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>