fstab: bump x64 tmpfs from 64 to 128 mb and split fstab files
x64 tmpfs was uncapped potentially taking up half of RAM while the
arm tmpfs was always capped to 64m. Some customers complained that
their programs write logs to the tmpfs filling up the RAM.

The decision was made to also cap x64 to 64mb but that breaks NXG<>CG
migration because opkg downloads ipk's to the tmpfs and the migration
images are over 64mb (the arm images fit into 64mb so leave it as is).

Bumping the x64 tpmfs to 128mb to fixes CG<>NXG migration and now we
also keep the fstab files separate: there's no need to append the arm
fstab from the recipe since the files have more than 1 divergence.

(CAR 681808 which prompted this work says it's ok to have both 64mb or
128mb tmpfs on x64).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>