migrate-nilrt: don't create -dev and -dbg packages
The ${PN} package contains only one shell script and is intended for a
feed conaining just the package with the script. The -dbg and -dev
packages are empty but pull in a lot of dependencies by default (make,
ncurses, etc).

Since nobody needs these two -dev & -dbg packages, don't create them
at all to have just the single useful package in the feed.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>