packagegroup-ni-base: move python dependencies to salt
The only thing in our distro running on target hardware having a hard
requirement on python is salt. We should not put python dependencies
directly in the base packagegroup as they should be added to salt
which needs them.

This does not change anything inside the image, it's merely cosmetic,
changing the way dependencies get pulled:

Instead of
packagegroup-base -> python3-*
we now have
packagegroup-base -> salt-minion -> salt-common -> python3-*

(python3-futures is added in the main salt recipe)

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>