nilrt-image-common: symlink /usr/bin/python to python3
By default the python 3 recipe does not create the symlink but salt
and other NILRT userspace tools expect it. The reason the python 3
recipe doesn't symlink by default is that python 2 does it and is the
default "python" in OE-core (even in side by side python 2/3 installs).

So, because we install only python 3 on targets, we need to create
it ourselves. More details in the code comments.

If someone installs python 2 from the feeds alongside
the default python 3, then v2 will take precedence (this is normal
because they asked for python 2).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>