nilrt-image-common: package the kernel for older NILRT
We keep removing the kernel from the standard /boot paths for backwards
compatibility reasons: each older NILRT image (safemode/runmode)/(arm/x64)
puts its kernel in some other place.

For ARM we package the kernel and DTB's in a custom path from where
nibuild/os-common will gobble it up and create an itb file. The reason
why we don't create the itb directly here in OE is that os-common
unpacks the safemode image and hacks up the filesystem before
re-packing it / creating the itb. We'd depend on os-common which
depends on the OE image we produce.

In the future I plan to move the ARM safemode filesystem hacking logic
to OE so we can also generate the itb in a cleaner way like we do for
NXG, but currently this work is outside the scope for this change.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>