niconsole-image-safe: Remove radeon firmware from safemode
The reason I'm removing it here and not at an upper layer in the image
hierarchies via the PACKAGE_REMOVE var from nilrt-images-common.inc is
that the only thing we remove there is the CG kernel & modules. Since
the CG kernel & modules will be built in OE shortly, that entire logic
will be deleted soon, so it will be better to keep this here untouched.

It also just makes more sense to remove stuff from the safemode image
in the safemode image recipe, not in something being inherited in all
images :)

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>