packagegroup-ni-restoremode: cleanup recipe
by moving the kernel modules to RRECOMMENDS to avoid build races as
this packagegroup doesn't install linux-nilrt (so we don't know if the
module ipks are built or not) and remove the DEPENDS on linux-nilrt.

Also remove the stale comment: this should have been removed in an
earlier commit but was overlooked.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>