xilinx-zynq(hf): move & rename device code var from NXG distro
We need this variable when generating itb's and dtb's for both older
and newer NILRT now that the kernel will be moved to OE for older
NILRT, so put it in xilinx-zynqhf.conf (included by xilinx-zynq.conf).

Rename it to reflect the more broad usage than migration and
deduplicate the list.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>