linux-nilrt: bump 4.9 -> 4.14
And also rename the recipes to _git to follow the OE convention for
git based recipes (PV is set inside the recipe from the git branch
data, the version in the name is useless and only triggers renames
when upgrading).

x64 now reuqires an elfutils-native dependency for gelf.h which the
kernel depends on.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>