module-tools: sync versioning-utils with the p4 version
Apparently this recipe was not being used for anything so I'm just
moving the remaining logic from p4 (added in the last 4 years) and
using it the same way it was used in older NILRT versions.

The names and path for the squashfs'es differ between older NLIRT and
this unused NXG one: I've chosen to go with the older name of
"tools-squashfs-image.squashfs" in the "tools" subdir, even though the
NXG name is better, because there might be client teams depending on
the name and I don't want to risk unnecessary breakage.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>