images: correctly build the two "older NILRT" squashfs images
We need two squashfs images for older NILRT now that the kernel build
has moved to OE/git instead of nibuild/p4:

1. A toolchain image containing gcc & others: module-tools.bb
2. An image containing kernel headers necessary to build modules:
module-headers.bb

Create these and delete the useless linux-nilrt-squashfs.inc logic
which created a kernel package containing a squashfs: It was not
used anywhere and most of it was dead code and not the right approach
since it was manually creating the squashfs instead of using
IMAGE_FSTYPES. The important bits from linux-nilrt-squashfs.inc have
moved to linux-nilrt.inc (the actual kernel headers packaging).

I verified the two produced squashfs contents are identical to the
ones exported by nibuild's kernel & module_tools components.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>