images: older nilrt: package the module verisioning squashfs'es
We need the older nilrt runmode images to contain squashfs'es
necessary for on-target module compilation and versioning.

This means we need to add the squashfs'es to the two runmode images:
nixfce-image.bb and niconsole-image.bb. The former is used by x64 with
embedded UI hence the name "xfce" and the other is used by ARM with no
embedded UI hence the name "console" :)

I know the naming of the images is confusing and there are so many
images between older and NXG NILRTs.. I hope this will be fixed soon
because these images have significant overlaps, the whole thing is in
need of a refactoring / re-naming.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>