linux-nilrt: unify header export logic
Now we're building the same kernel in the same place for both older
and newer NILRT (NXG) so it doesn't make sense to have two independent
scripts to package the kernel headers for older and newer NILRT.

We still package the headers separately: NXG headers are in kernel-dev
and older NILRT headers are in ${PN}-module-versioning-headers which
ends up in the headers squashfs image.

We'll remove the separate packaging and use kernel-dev for both
distros when client driver teams agree to move to kernel-dev from the
squashfs (probably next year).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>