grub-efi: compile our own efi binary again
We tried re-using the standard bootx64.efi binary compiled by the
normal OE build for our chainloaded rootfs grub but it doesn't work
because of the default prefix in the bootx64.efi binary which points
to ($root)/EFI/BOOT instead of the simple ($root)/boot which we need.

So compile our own binary with the custom path again.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>