module-tools: versioning_utils.sh: fix header squashfs mount
I completely botched the merge from p4 to git in the previous commit,
sorry about that :) This file needs to be identical to the p4 version,
so this commit re-adds the missing pieces to sync with the p4 version.

This script is not used on NXG so it's safe to discard all the
previous changes present in this repo (that version was initially
intended to be used on NXG in 2015 but the design changed since then).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>