packagegroup-ni-base: add squashfs kernel module on ARM
Since we've moved the older NILRT kernel build to OE, we must ensure
each module gets bundled in the correct image as opposed to dumping
all modules in each image as nibuild did.

The ARM images need the squashfs module to do on-target kernel module
versioning (kernel headers and tools are in squashfs'es).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>