linux-nilrt: properly package older NILRT kernel
On older NILRT we used to install the kernel in its custom paths via
ROOTFS_POSTPROCESS_COMMAND which means it was not properly packaged,
eg. on x64, it was outside the kernel-image-bzimage package.

To support multiple kernel installs from the extras feed (for eg a
lockdep enabled debug kernel in the runmode image) we need the kernel
to be properly packaged, so use the KERNEL_IMAGEDEST variable to move
it to a custom destination.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>