linux-nilrt.inc: Disable dpm for radeon video driver module
The dynamic power management (dpm) feature of the radeon video driver
is broken in kernels 4.13 or later. The issue manifests itself as a
white screen as soon as the console framebuffer switches to using the
radeon driver. This is only a problem with text consoles i.e. X works
as expected. The problem was observed on PXIe-8880 with a Advanced
Micro Devices, Inc. Seymour [Radeon E6460] video chip.

Turn off dpm via module parameters. As a side benefit this will also
improve the real-time characteristics of this driver.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Natinst-CAR-ID: 714692
Natinst-ReviewBoard-ID: 261066