python2: don't symlink by default to /usr/bin/python
OE has the policy of python 2 being the target default so every time
it gets installed it symlinks itself as the default python even if v3
is already installed (clobbering the v3 symlink).

On NILRT, this breaks Salt and multiple client teams which expect v3
to always be the default.

Since OE is planning to make python 3 the default in the near future
and significantly rework the "python" recipe which is currently v2,
just remove the hardcoded symlink from python-core.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>