python3-salttesting: remove bbappend
We do not have any changes in this recipe/repo because everything is
upstream. The only reason we had this bbappend is because we needed a
newer version than what the recipe in meta-cloud-services provides.

So we remove this bbappend version bump so we can do a real
meta-cloud-services recipe bump which we can also upstream, thus
avoiding the need to point to internal NI repos.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>
Natinst-CAR-ID: 725258