base-files: change /var/volatile to %-based size
Previously, the size of /var/volatile was set to a fixed value in fstab: 64MB
for ARM and 128MB for x64 targets.  These values were set based on the smallest
memory targets.  These small fixed values cause problems on larger memory
targets when doing large opkg installs because the ipks get downloaded to
/var/volatile.

To improve this situation, change to using a percent based size so that the
size will scale with available RAM.

To determine the percentage, we looked at the total physical memory reported
on the smallest memory ARM target which was 250160KB.  In order to ensure at
least the current /var/volatile size of 64MB, the percentage should be 26%.

Signed-off-by: Ken Sharp <ken.sharp@ni.com>
Acked-by: Alex Stewart <alex.stewart@ni.com>
Natinst-CAR-ID: 728695
Natinst-ReviewBoard-ID: 277314