ninextos: Move/rename install folder for provisioning image
The opkg utility did not like installing <dot>newNILinuxRT folder
to the root directory, so renamed the folder to newNILinuxRT since
it only exists on the target for a few minutes while provisioning
is taking place.  Also updated references in bootimage.cfg to
reflect new location and name.  Removed ARM support as well.

Signed-off-by: Bill Pittman <bill.pittman@ni.com>
Acked-by: Haris Okanovic <haris.okanovic@ni.com>
Acked-by: Alex Stewart <alex.stewart@ni.com>
Natinst-ReviewBoard-ID: 311755