ni_provisioning: Add PROVISION_BOOT_SCHEME
An option to toggle between different methods of booting the OS.

"efi-ab" configures the machine's EFI firmware to boot redundant A/B
images (ESPs). This is required for NILRT power-loss safe image updates
via RAUC and is the default.

"efi-auto" configures a single ESP with a default /efi/boot/bootx64.efi
boot loader such that it can be automatically discovered and booted by
firmware. This option is useful for booting removable storage and VMs
where EFI nvram configuration cannot persist.

Signed-off-by: Haris Okanovic <haris.okanovic@ni.com>
Acked-by: Bill Pittman <bill.pittman@ni.com>
Acked-by: Alex Stewart <alex.stewart@ni.com>
Natinst-ReviewBoard-ID: 310313