nilrt-nxg: Create packages to install nilrt-nxg
nilrt-nxg-gateway should only install on a safemode compatible
system and then migrate that system to the new A/B image system.
nilrt-nxg-base should only be installed on a new system image
and will install the current DISTRO_VERSION of the minimal image
and configure the boot partition.  Both recipes install a nilrt-install
script to install the payload to the appropriate partition.
Acked-by: Haris Okanovic <haris.okanovic@ni.com>
Acked-by: Alex Stewart <alex.stewart@ni.com>
Natinst-ReviewBoard-ID: 311755