base-files: Use full path to runlevel command in safemode-ps1.sh
The changes to safemode-ps1.sh made in aeccc9be (base-files:
Prepend "safemode" to PS1 when logged into rc4) applied to
all users but runlevel is a system binary and not on
non-admin user paths.

This change ensures the script uses the full path to runlevel
(/sbin/runlevel) such that the script works for all users.

Signed-off-by: Charlie Johnston <charlie.johnston@ni.com>
Acked-by: Haris Okanovic <haris.okanovic@ni.com>
Acked-by: Alex Stewart <alex.stewart@ni.com>
Natinst-AZDO-ID: 571756
Natinst-ReviewBoard-ID: 322058