niacctbase: add lvuser to video group
'lvuser' is not currently part of the 'video' group. The 'video' group
is used to control access to: video capture devices, 2D/3D hardware
acceleration, and framebuffer devices.

This results in X errors like:

     libGL error: failed to open drm device: Permission denied
     libGL error: failed to load driver: r600

Add 'lvuser' to the 'video' group to avoid these errors. This is
standard practice in other Linux distros.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Haris Okanovic <haris.okanovic@ni.com>
Acked-by: Alex Stewart <alex.stewart@ni.com>
Natinst-AZDO-ID: 206148
Natinst-ReviewBoard-ID: 325144