util-linux: fixup CLASSOVERRIDE DEPENDS
Specifying DEPENDS_class-target_appends (or variations otherof) results
in the internal value of DEPENDS from util-linux.bb being overwritten
entirely with the DEPENDS_class-target value. As a result, `libcap-ng`
is never added to the recipe sysroot and the util-linux:do_configure
task fails with an error that it cannot find libcap-ng.

Instead, use the syntax: `DEPENDS_append_class-target` to have bitbake
append to the `DEPENDS` variable in a non-native recipe build. This is
the accepted syntax in OE upstream.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>