busybox: disable the MOUNT_NFS feature
The busybox feature CONFIG_FEATURE_MOUNT_NFS is only necessary in linux
kernels perior to 2.6. Further, this feature requires the rpc.h headers
from glibc<2.26, which have since been deprecated. As a result, busybox
upstream's recommendation is to disable this feature on newer kernels.

http://lists.busybox.net/pipermail/busybox/2017-September/085848.html

OE upstream has done the same in their busybox defconfig.

https://www.openembedded.org/pipermail/openembedded-core/2015-November/113155.html

Disable it in our defconfig for the same reasons.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>