glibc: fix EAGAIN retry loop in pthread_cond_broadcast()
Commit ac38b649bcb9 ("glibc: pi-condvars: EAGAIN retry in
pthread_cond_ broadcast() and signal()") introduced a retry loop for
an EAGAIN error on a FUTEX_CMP_REQUEUE_PI futex syscall in
pthread_cond_broadcast(). However the futex value is not re-read
before re-invoking the futex syscall on retry. This can result in an
infinite retry loop if the reason the kernel returned EAGAIN is
because the futex value changed after the private lock was released.

Update the futex value and other associated sequence counters before
retrying the syscall.

Fixes: ac38b649bcb9 ("glibc: pi-condvars: EAGAIN retry in pthread_cond_ broadcast() and signal()")
Natinst-AZDO-ID: 1024009
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>