tpm2: Remove ni/tpm2 support in Dunfell.
tpm2 recipes have been updated to newer versions in upstream Dunfell
meta-measured. Consequently, ni/tpm2 forks have to be rebased on to
the newer versions. This work is significant and will be done at a
later date. Until then, removing support for ni/tpm2 in oe. Upstream
tpm2 packages are still available in the extras feed.

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>