coreutils: Fix warnings due to upstream updates.
Change priority of packages to a value higher than the priority of
coreutils. This fixes a warning that the binaries have multiple
providers with the same priority.
Update RDEPENDS variable name to fix warning that this key replaces
the original RDEPENDS_coreutils key in coreutils instead of
appending to it.

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>