linux, images: Only rename the versioned kernel file on image builds
Standard OE builds generate a versioned kernel image file on build
which is then symlinked to e.g. 'bzImage-<ver>' with a 'bzImage'
symlink pointing to it.

For backwards compatibility the current implementation renames the
versioned kernel image file into an unversioned file e.g.
bzImage-<ver> becomes bzImage. This also in essence overwrites the
existing symlink. This is done for all NILRT kernels via a
do_install_append() snippet in the common nilrt kernel
'linux-nilrt.inc'. The reason behind it is to accommodate a fair
amount of legacy post processing and tooling that happens outside OE.

However this renaming is undesirable for additional kernel builds that
are not the main kernel included in the image. For example for kernels
installed from package feeds we do want the versioned kernel image +
symlink behavior. This allows us to install multiple kernels side by
side without the files conflicting. It also allows us to switch
kernels by switching the associated symlink.

Move the kernel renaming logic into the image build recipes as opposed
to kernel recipes since the naming scheme is an image level
implementation decision. This allows us to keep the versioned kernel +
symlink behavior for extra kernels installed via the package feed.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>