linux-nilrt: Handle kernel installs/uninstalls from packages
Add package maintainer scripts to handle a kernel install/uninstall
from package feeds on a NILRT distribution.

The current boot flow expects the kernel at
${KERNEL_IMAGEDEST}/${KERNEL_IMAGETYPE} (e.g. /boot/runmode/bzImage)
and it's currently installed as a plain file in the built images.

On install we move the existing kernel out of the way and we record
its version in a 'kernel-versions' file which keeps track of available
kernels. The new kernel image gets installed as a versioned bzImage
file and we create a symlink to point to it.

On uninstall, in 'prerm', before removing anything we run sanity
checks to confirm we have a kernel to fall back to. If the 'prerm'
checks succeed we update the symlink (if needed) to the fallback
kernel and clean-up the 'kernel-versions' file.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>