libglade: move recipe from meta-openembedded.
Instead of carrying this patch forward every rebase in
meta-openembedded, move recipe to meta-nilrt.

commit 9dcf44f ("libglade: remove") dropped libglade because it
was broken by recipe specific sysroots (glib-2.0-native was missing).

Though it is old/deprecated/unmaintained, libglade is useful because
it's a dependency for the florence virtual keyboard which is the
smallest virtual keyboard in terms of disk space / memory usage.

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>