Add nsswitch.conf that queries the ldap database.
This updated nsswitch.conf file was originally provided by the
meta-cloud-services layer. The layer had a bug that enabled
installation of this file even when OpenLDAP was not enabled as a
DISTRO_FEATURE. This bug was fixed in the Dunfell branch and the file
is no longer added to the image because we do not enable OpenLDAP.
We still want to keep this version of the file to avoid breaking
customers who might expect this behavior after installing ldap
libraries.

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>