nisdbootconfig: add a new recipe for linux runmode to configure boot from sd card
Artemis uses SD card as its main disk instead of NAND drive. Some configurations
are required for it to boot properly. This includes the mountpoints for the sd
card partitions and the fw_printenv confguration. The nisdbootconfig script needs
to run before mountall is executed, hence this recipe will install and update the
script at run level S with 00 priority.

This commit includes adding a new recipe "nisdbootconfig" which contains the
nisdbootconfig script and the bitbake recipe.

Signed-off-by: wkoe <wilson.koe@ni.com>