nisdbootconfig: Fixup commit 801d331 breaking x64 images.
Commit 801d331 added nisdbootconfig to x64 images too. The recipe
inherits from update-rc.d but does not install any file to
/etc/init.d on x64 targets. This causes the post-inst of the package
to fail on x64 targets when the following command is run:
update-rc.d $OPT nisdbootconfig start 00 S .
because the nisdbootconfig file does not exist on that image.

To fix this, add the package to only ARM images in
packagegroup-ni-base since the package is common to both safemode and
runmode.

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>