nilrt-base.inc: install core packages from feeds
All packages which must be installed from feeds should be declared via
the _NODEPS subvariable. Otherwise, bitbake is allowed to rebuild the
recipe in the local workspace (rather than getting it from the feed).

Enforce this rule on the nilrt-base.inc so that all NILRT images must be
configured with a remote IPK feed location in order to build.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>