nilrt.inc: Add the lv2020 feed to the feed URIs list
The lv2020 feed hosts the ni-rtlog package that is required to be
installed in the base image. The previous solution moved the ni-rtlog
to the ni-main feed, but that broke LabVIEW installations for a
number of downstream clients who were replacing the ni-main feed with a
local feed.
To workaround the breakage, install ni-rtlog from the lv2020 feed
instead. The previous solution can be adopted again (reverting this
change) when downstream clients are able to test with the new base
image that includes the ni-rtlog package.

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>