recipes-ni: add ni-hw-support packages
Create a new family of packages - ni-hw-support* - which deliver support
scripts (initscripts) and tools needed specifically to support NI
product line hardware.

The first installment is the ni-rename-ifaces initscript, which has been
migrated (without modification) from os-common. It will be delivered via
the ni-hw-support-common package, since it is common to all product
lines.

Add the family's meta-package - ni-hw-support - to the base
packagegroup, so that it is installed on all base system images.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>