packagegroup-ni-base, busybox: Add dependency on busybox-udhcpd
Commit 3221e817ebc2 ("busybox-udhcpd: Move wireless udhcpd
configuration file from os-common") adds a configuration file to the
busybox-udhcpd package.

This configuration file has to be included in the base NILRT image.
Add it to the corresponding package group.

Additionally the default busybox-udhcpd recipe installs an init script
which is not used in NILRT images. On these images udhcpd is invoked
directly from ifplugd action scripts. Remove it.

Suggested-by: Chaitanya Vadrevu <chaitanya.vadrevu@ni.com>
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>