conf: move local.conf changes to distro
To reduce the complexity of the local.conf, move settings which are
frequently set by build machines or which should not be set by the user
at all, into other configuration files.

In this commit, distro.conf takes ownership of:
* the packaging system (via PACKAGE_CLASSES)
* the appropriate image recipes to keep via (rm_work)
* the default MACHINE value (x64)
* the useradd-staticids configuration

Other local.conf settings have been moved to the `auto.conf` and
`ni-org.conf` files in nilrt.git.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>
Rebased-by: Alex Stewart <alex.stewart@ni.com>