nilrt-base-bundle-image: fixup bzImage in bundle
The UEFI spec requires vfs types which do not support links (vfat, fat,
ISO, etc). Therefore, RAUC has to extract the bundle image into a
FAT-formatted niboot partition. Before we do that, we need to eliminate
the bzImage symlink, so move the versioned kernel image overtop the
generic symlink.

* The new move operation also uses the new bzImage location (`/boot`,
  rather than `/boot/runmode`) for the RAUC images.

* Also stop trying to remove the `/boot/runmode` directory, since it no
  longer exists.

We cannot use the `KERNEL_IMAGEDEST` variable to determine the bzImage
location, because it was previously being provided by the distro config
file, but is no longer. The kernel recipes inherit it from the
kernel.bbclass now, so they have it at kernel-build-time; but the bundle
image recipe has to do without.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>