dist-nilrt-efi-ab-gateway: Disable -lic package creation
The recipe as is creates a separate accompanying -lic package and
adds a "RECOMMENDS" to it from the base dist package.
When we publish packages to ni.com, we do not want to publish a
separate -lic package for each dist package because that can grow out
of control very quickly and start to pollute the feed.
So, disable the -lic package creation and install the license file
as part of the base dist package.