nilrt.conf: rename the release to "unstable"
Ever-changing planning requirements mean that it is difficult to assert
a version number on a purely development branch like 'dunfell'. Instead
of trying to guess the next labview release version, use a static string
like 'unstable' as the release name.

As a NILRT branch ref falls onto the commitment schedule, its release
name can be changed to a version number.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>