conf/local.conf: reconcile local.conf with dev-mainline
Manually reconcile the state of the local.conf with what we have in the
dev-mainline (nilrt/master/dunfell). The heads have deviated enough that
this is the best option for merging them.

Also remove the distinction between the `nilrt` and `nilrt-nxg` distro
versions, since the latter is no longer supported in this mainline.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>