migrate org-specific recipes from meta-orgconf
The meta-orgconf layer is being deprecated. Unfortunately, there are
still some recipes which rely on org-specific information to build -
particularly the recovery media.

Move the required classes and recipes from meta-orgconf into meta-nilrt,
so that we can deprecate the former.

This commit does not modify the contents of the moved files.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>