recipes-ni: deprecate zynq-safemode-cfgs
I think the zynq-safemode-cfg recipes were necessary at some point in
the past, as a way to enable feed-based migration of ARM targets. That
feature has been deprecated for years, and I can find no reference to
any component which builds or uses these IPKs.

Deprecate the recipes and remove them from the layer, for cleanliness.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>