kernel: Add NO_HZ_FULL kernel flavor packages and package group
In certain types of real-time applications it is desirable to reserve CPU
cores for real-time threads and avoid interference from the kernel as much
as possible. This includes things like the kernel scheduler tick which can
create jitter spikes.

Provide packages for a kernel configured with NO_HZ_FULL. This allows users
with strict RT requirements to install a kernel that will adaptively try to
shutdown the scheduler thick whenever possible. This comes at the expense
of some overhead in user <-> kernel transitions (i.e. syscalls,
exceptions, and interrupts).

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>