virtual/kernel: Relocate git source/branch related information
The current implementation assumes all current kernel flavors are built
from the same source and as such the git repo and branch information is
stored in the common include file. To allow for future expansion and adding
more kernel flavors relocate this information to the corresponding recipe.
This allows for finer grain control on the version used for different NILRT
kernel flavors.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>