niwatchdogpet: move watchdogpet later to fix fw_printenv read error
While developing safemode for artemis, we found out that fw_printenw call in
niwatchdogpet cannot function properly because the config lives in
/boot/uboot/uboot.env, which requires the /boot partition to be mounted first.
The mounting of /boot for artemis depends on /dev/mmcblk0p1 node to be ready,
which requires udev to be initialized first. Hence, by moving niwatchdogpet
from S04 to S05, we can resolve this issue.

This commit changes niwatchdogpet startup from S04 to S05.

Signed-off-by: wkoe <wilson.koe@ni.com>