pkggrp-ni-desirable: add ni-grpc-device
Add ni-grpc-device to the desirable packagegroup, so that it is always
built and included in the package feeds. Ideally, it would be in the
core feed, but the current architecture of the sumo pipeline makes that
prohibitively expensive to implement at the moment. So instead, accept
that it will be in the extra/ feed for now.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>