ni-grpc-device: fix patches for upstream builds
The upstream grpc-device repo uses GH workflows to verify that patches
do not break compilation on native Ubuntu and Windows build
environments, and using the legacy NILRT build toolchain. The current
patchset is not upstreamable, because it breaks Windows and legacy NILRT
compilation.

Fixup the patches to resolve upstream issues and make patches 0001-0004
upstreamable.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>