nilrt-image-common: ensure python3 stays out of base images
The base images should contain the bare minimum set of packages for
all NI use cases, and some of those use cases are very tightly
constrained, such as ARM safemode images. The python3 runtime is large
and it keeps finding its way accidentally into places, so add a
specific PACKAGE_EXCLUDE check for it to alert us if it gets added
as a dependency without us realizing.

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>