kernel-devsrc: add elfutils-dev as dependency
The dkms module rebuild machinery expects to have objtool if elfutils-dev
is on the system at the time of rebuild, but if elfutils-dev was not
present at the time that kernel-devsrc was installed, then objtool is not
generated. Resolve this by making elfutils-dev a dependency; thus it
should always be present if kernel-devsrc is.

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
Natinst-AzDO-ID: 1470108
(cherry picked from commit 92ff3a34f5b2222106a4fc94a035f76e2e500b72)